requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
asgiref==3.5.1
    # via django
autopep8==1.6.0
    # via -r requirements.in
backports-zoneinfo==0.2.1
    # via django
django==4.0.4
    # via
    #   -r requirements.in
    #   djangorestframework
django-dotenv==1.4.2
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
numpy==1.22.3
    # via pandas
pandas==1.4.2
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
pycodestyle==2.8.0
    # via autopep8
python-dateutil==2.8.2
    # via pandas
pytz==2022.1
    # via
    #   djangorestframework
    #   pandas
six==1.16.0
    # via python-dateutil
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via autopep8
